+++ packages/react-ui/storybook/stories/widgetsUI/FormulaWidgetUI.stories.js
@@ -27,27 +27,30 @@ const Template = (args) => <FormulaWidgetUI {...args} />;
 const data = 10000;
 
 export const Empty = Template.bind({});
-Empty.args = {};
+
+export const Basic = Template.bind({});
+const BasicProps = { data };
+Basic.args = BasicProps;
 
 export const Text = Template.bind({});
 const TextProps = { data: `$${data}` };
 Text.args = TextProps;
 
-export const ValueUnit = Template.bind({});
-const ValueUnitProps = { data: { value: data, unit: '$' } };
-ValueUnit.args = ValueUnitProps;
+export const ValuePrefix = Template.bind({});
+const ValuePrefixProps = { data: { value: data, prefix: '$' } };
+ValuePrefix.args = ValuePrefixProps;
+
+export const ValueSuffix = Template.bind({});
+const ValueSuffixProps = { data: { value: data, suffix: '€' } };
+ValueSuffix.args = ValueSuffixProps;
 
 export const FormatterText = Template.bind({});
-const FormatterTextProps = { data, formatter: (v) => `$${v}` };
+const formatter = (value) =>
+  Intl.NumberFormat('en', {
+    maximumFractionDigits: 1,
+    minimumFractionDigits: 0,
+    notation: 'compact',
+    compactDisplay: 'short'
+  }).format(value);
+const FormatterTextProps = { data, formatter };
 FormatterText.args = FormatterTextProps;
-
-export const FormatterValueUnit = Template.bind({});
-const FormatterValueUnitProps = { data };
-FormatterValueUnit.args = FormatterValueUnitProps;
-
-export const FormatterValueUnitBefore = Template.bind({});
-const FormatterValueUnitBeforeProps = {
-  data,
-  unitBefore: true
-};
-FormatterValueUnitBefore.args = FormatterValueUnitBeforeProps;